+++ android/src/main/kotlin/com/hiennv/flutter_callkit_incoming/FlutterCallkitIncomingPlugin.kt
@@ -1,10 +1,12 @@
 package com.hiennv.flutter_callkit_incoming
 
+import android.annotation.SuppressLint
 import android.app.Activity
 import android.content.Context
 import android.os.Handler
 import android.os.Looper
 import androidx.annotation.NonNull
+import androidx.annotation.Nullable
 
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.embedding.engine.plugins.activity.ActivityAware
@@ -20,11 +22,33 @@ class FlutterCallkitIncomingPlugin : FlutterPlugin, MethodCallHandler, ActivityA
 
     companion object {
 
+        @SuppressLint("StaticFieldLeak")
+        private var instance: FlutterCallkitIncomingPlugin? = null
+
+        public fun getInstance(): FlutterCallkitIncomingPlugin  {
+            if(instance == null){
+                instance = FlutterCallkitIncomingPlugin()
+            }
+            return instance!!
+        }
 
         private val eventHandler = EventCallbackHandler()
 
         fun sendEvent(event: String, body: Map<String, Any>) {
             eventHandler.send(event, body)
+        }
+
+        private fun sharePluginWithRegister(@NonNull flutterPluginBinding: FlutterPlugin.FlutterPluginBinding, @Nullable handler: MethodCallHandler) {
+            if(instance == null) {
+                instance = FlutterCallkitIncomingPlugin()
+            }
+            instance!!.context = flutterPluginBinding.applicationContext
+            instance!!.callkitNotificationManager = CallkitNotificationManager(flutterPluginBinding.applicationContext)
+            instance!!.channel = MethodChannel(flutterPluginBinding.binaryMessenger, "flutter_callkit_incoming")
+            instance!!.channel?.setMethodCallHandler(handler)
+            instance!!.events =
+                EventChannel(flutterPluginBinding.binaryMessenger, "flutter_callkit_incoming_events")
+            instance!!.events?.setStreamHandler(eventHandler)
         }
 
     }
@@ -33,34 +57,79 @@ class FlutterCallkitIncomingPlugin : FlutterPlugin, MethodCallHandler, ActivityA
     ///
     /// This local reference serves to register the plugin with the Flutter Engine and unregister it
     /// when the Flutter Engine is detached from the Activity
-    private lateinit var activity: Activity
-    private lateinit var context: Context
-    private lateinit var callkitNotificationManager: CallkitNotificationManager
-    private lateinit var channel: MethodChannel
-    private lateinit var events: EventChannel
+    private var activity: Activity? = null
+    private var context: Context? = null
+    private var callkitNotificationManager: CallkitNotificationManager? = null
+    private var channel: MethodChannel? = null
+    private var events: EventChannel? = null
+
 
     override fun onAttachedToEngine(@NonNull flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
         this.context = flutterPluginBinding.applicationContext
-        callkitNotificationManager = CallkitNotificationManager(this.context)
+        callkitNotificationManager = CallkitNotificationManager(flutterPluginBinding.applicationContext)
         channel = MethodChannel(flutterPluginBinding.binaryMessenger, "flutter_callkit_incoming")
-        channel.setMethodCallHandler(this)
+        channel?.setMethodCallHandler(this)
         events =
             EventChannel(flutterPluginBinding.binaryMessenger, "flutter_callkit_incoming_events")
-        events.setStreamHandler(eventHandler)
+        events?.setStreamHandler(eventHandler)
+ //       sharePluginWithRegister(flutterPluginBinding, this)
     }
+
+    public fun showIncomingNotification(data: Data) {
+        data.from = "notification"
+        callkitNotificationManager?.showIncomingNotification(data.toBundle())
+        //send BroadcastReceiver
+        context?.sendBroadcast(
+            CallkitIncomingBroadcastReceiver.getIntentIncoming(
+                requireNotNull(context),
+                data.toBundle()
+            )
+        )
+    }
+
+    public fun startCall(data: Data) {
+        context?.sendBroadcast(
+            CallkitIncomingBroadcastReceiver.getIntentStart(
+                requireNotNull(context),
+                data.toBundle()
+            )
+        )
+    }
+
+    public fun endCall(data: Data) {
+        context?.sendBroadcast(
+            CallkitIncomingBroadcastReceiver.getIntentEnded(
+                requireNotNull(context),
+                data.toBundle()
+            )
+        )
+    }
+
+    public fun endAllCalls() {
+        val calls = getDataActiveCalls(context)
+        calls.forEach {
+            context?.sendBroadcast(
+                CallkitIncomingBroadcastReceiver.getIntentEnded(
+                    requireNotNull(context),
+                    it.toBundle()
+                )
+            )
+        }
+        removeAllCalls(context)
+    }
+
 
     override fun onMethodCall(@NonNull call: MethodCall, @NonNull result: Result) {
         try {
-            callkitNotificationManager = CallkitNotificationManager(this.context)
             when (call.method) {
                 "showCallkitIncoming" -> {
                     val data = Data(call.arguments())
                     data.from = "notification"
-                    callkitNotificationManager.showIncomingNotification(data.toBundle())
+                    callkitNotificationManager?.showIncomingNotification(data.toBundle())
                     //send BroadcastReceiver
-                    context.sendBroadcast(
+                    context?.sendBroadcast(
                         CallkitIncomingBroadcastReceiver.getIntentIncoming(
-                            context,
+                            requireNotNull(context),
                             data.toBundle()
                         )
                     )
@@ -68,9 +137,9 @@ class FlutterCallkitIncomingPlugin : FlutterPlugin, MethodCallHandler, ActivityA
                 }
                 "startCall" -> {
                     val data = Data(call.arguments())
-                    context.sendBroadcast(
+                    context?.sendBroadcast(
                         CallkitIncomingBroadcastReceiver.getIntentStart(
-                            context,
+                            requireNotNull(context),
                             data.toBundle()
                         )
                     )
@@ -78,9 +147,9 @@ class FlutterCallkitIncomingPlugin : FlutterPlugin, MethodCallHandler, ActivityA
                 }
                 "endCall" -> {
                     val data = Data(call.arguments())
-                    context.sendBroadcast(
+                    context?.sendBroadcast(
                         CallkitIncomingBroadcastReceiver.getIntentEnded(
-                            context,
+                            requireNotNull(context),
                             data.toBundle()
                         )
                     )
@@ -89,9 +158,9 @@ class FlutterCallkitIncomingPlugin : FlutterPlugin, MethodCallHandler, ActivityA
                 "endAllCalls" -> {
                     val calls = getDataActiveCalls(context)
                     calls.forEach {
-                        context.sendBroadcast(
+                        context?.sendBroadcast(
                             CallkitIncomingBroadcastReceiver.getIntentEnded(
-                                context,
+                                requireNotNull(context),
                                 it.toBundle()
                             )
                         )
@@ -112,12 +181,13 @@ class FlutterCallkitIncomingPlugin : FlutterPlugin, MethodCallHandler, ActivityA
     }
 
     override fun onDetachedFromEngine(@NonNull binding: FlutterPlugin.FlutterPluginBinding) {
-        channel.setMethodCallHandler(null)
+        channel?.setMethodCallHandler(null)
     }
 
 
     override fun onAttachedToActivity(binding: ActivityPluginBinding) {
         this.activity = binding.activity
+        this.context = binding.activity.applicationContext
     }
 
     override fun onDetachedFromActivityForConfigChanges() {
@@ -125,6 +195,7 @@ class FlutterCallkitIncomingPlugin : FlutterPlugin, MethodCallHandler, ActivityA
 
     override fun onReattachedToActivityForConfigChanges(binding: ActivityPluginBinding) {
         this.activity = binding.activity
+        this.context = binding.activity.applicationContext
     }
 
     override fun onDetachedFromActivity() {}