+++ example/lib/main.dart
@@ -161,7 +161,7 @@ class _MyAppState extends State<MyApp> {
   }
 
   Future<void> makeFakeCallInComing() async {
-    await Future.delayed(const Duration(seconds: 7), () async {
+    await Future.delayed(const Duration(seconds: 10), () async {
       this._currentUuid = _uuid.v4();
       var params = <String, dynamic>{
         'id': _currentUuid,